backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.24.0
pandas>=2.1.3
numpy>=1.26.2
python-multipart>=0.0.6
pydantic>=2.6.0
scikit-learn
umap-learn
openai>=1.50.0
networkx==3.6
